buildwithai/requirements.txt
asgiref==3.7.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
Django==3.2.25
djangorestframework==3.15.1
drf-yasg==1.21.10
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.1
googleapis-common-protos==1.70.0
grpcio==1.62.3
grpcio-status==1.62.3
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
packaging==24.0
proto-plus==1.26.1
protobuf==4.24.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pytz==2025.2
PyYAML==6.0.1
requests==2.31.0
rsa==4.9.1
simplejson==3.20.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.7
